LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/merge-timeline-offset-keyframes-expected.txt


PASS Keyframes with same easing and timeline offset are merged.
FAIL Keyframes with same timeline offset but different easing function are not merged. assert_equals: number of frames:  expected 5 but got 4
